finalize schema awareness
