+++ python/src/gpx2db/sql/proximity-calc/0250_insert_enriched_points.sql
@@ -5,7 +5,7 @@ insert into {schema}.newpoints (
   wkb_geometry
 )
 SELECT
-  nextval('newpoints_seq'),
+  nextval('{schema}.newpoints_seq'),
   segment_id,
   track_id,
   wkb_geometry
