+++ python/src/gpx2db/sql/proximity-calc/3400_view_count_circle_freq_all.sql
@@ -5,7 +5,7 @@ SELECT
   sum(num) as freq
 from (
   select *
-  from count_ls
+  from {schema}.count_ls
   UNION ALL
   select *
   from {schema}.count_ml_consecutive
