+++ python/src/gpx2db/sql/proximity-calc/4300_calc_categories.sql
@@ -4,21 +4,6 @@
 --- B) Enumerate all points and identify segments by looking for consecutive 
 ----   lines in same track, with same frequency
 
---- Create table
-drop table if exists {schema}.frequency_lines;
-create table {schema}.frequency_lines
-(
-    point_id_start integer not NULL,
-    point_id_end integer not NULL,
-    track_id integer not null,
-    segment_id integer not null,
-    freq integer not null,
-    category_segment integer,
-    constraint freq_lines_pk primary key (point_id_start )
-)
-tablespace pg_default;
-create index on {schema}.frequency_lines(point_id_end);
-
 -- first population
 insert into {schema}.frequency_lines (
 with fr as (
@@ -61,11 +46,35 @@ case when
     (f1.track_id != f2.track_id) or 
     (f1.segment_id != f2.segment_id) or 
     (f2.point_id_start is null) 
-then nextval('category_segment_seq')
-else currval('category_segment_seq') END as inc
+then nextval('{schema}.category_segment_seq')
+else currval('{schema}.category_segment_seq') END as inc
 from {schema}.frequency_lines f1
-left join frequency_lines f2 on f2.point_id_end = f1.point_id_start 
+left join {schema}.frequency_lines f2 on f2.point_id_end = f1.point_id_start 
 where tf.point_id_start = f1.point_id_start
 ) ;
 
+-- Create strings for each segment and insert it into table
+
+-- Dieses Art die Linestrings zu erzeugen ist nicht ganz zuverlässig. Eigentlich sollten nur Linestrings entstehen, aber ST_Linemerge schafft es in seltenen fällen nur einen MultiLineString zu erzeugen. Eigentlich sollte man nicht ST_Linemerge nutzen, sondern über die category_segment id's iterieren und mit ST_MakeLine alle vokommenden Punkte in der gegebenen Reihenfolge verketten.
+
+insert into {schema}.track_segment_freq_categories (
+select
+    fl.category_segment,
+    fl.freq,
+    null,
+    ST_Multi(ST_Linemerge(ST_Collect(ST_MakeLine(np1.wkb_geometry, np2.wkb_geometry))))
+from
+	{schema}.frequency_lines fl,
+    {schema}.newpoints np1,
+    {schema}.newpoints np2
+where
+	np1.point_id = fl.point_id_start  and
+    np2.point_id = fl.point_id_end
+group by fl.category_segment, fl.freq
+order by fl.category_segment, fl.freq
+);
+commit;
+-- insert logarithmic scale for 5 categories
+update  {schema}.track_segment_freq_categories
+set category = round(log((select power(max(freq), 1./4) from {schema}.track_segment_freq_categories), freq));
